Rename 'pages' app to 'hci'
