+++ src/config/urls.py
@@ -22,7 +22,7 @@ from django.urls import path
 
 urlpatterns = [
     path("up/", include("up.urls")),
-    path("", include("pages.urls")),
+    path("", include("hci.urls")),
     path("admin/", admin.site.urls),
 ]
 if not settings.TESTING:
