+++ src/hci/templates/pages/new_allele_haplotype.html
