+++ src/hci/templates/pages/new_curation.html
