+++ src/hci/templates/pages/new_disease.html
