+++ src/hci/templates/pages/new_publication.html
